pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "talkrefine"
version = "0.1.0"
description = "Local voice input with LLM refinement - speak naturally, get polished text"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "swenyang"}
]
keywords = ["voice", "speech-to-text", "llm", "dictation", "asr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "funasr>=1.3.0",
    "modelscope",
    "pyaudio",
    "pyperclip",
    "pyautogui",
    "keyboard",
    "requests",
    "pystray",
    "Pillow",
    "pyyaml",
]

[project.optional-dependencies]
torch-cpu = ["torch", "torchaudio"]
torch-cuda = ["torch", "torchaudio"]
whisper = ["openai-whisper"]
win32 = ["pywin32"]
all = ["talkrefine[torch-cpu,whisper,win32]"]

[project.scripts]
talkrefine = "talkrefine.app:main"

[project.urls]
Homepage = "https://github.com/swenyang/talk-refine"

[tool.setuptools.packages.find]
include = ["talkrefine*"]
